+++ propnet/core/utils.py
@@ -4,16 +4,6 @@ from uuid import uuid4, uuid5, NAMESPACE_URL
 
 from habanero.cn import content_negotiation
 from monty.serialization import loadfn, dumpfn
-from monty.json import jsanitize
-
-from maggma.stores import MongoStore
-from maggma.builder import Builder
-from maggma.runner import Runner
-from propnet import logger
-from propnet.core.quantity import Quantity
-from propnet.core.materials import Material
-from propnet.core.graph import Graph
-from pydash import get
 
 
 NAMESPACE_PROPNET = uuid5(NAMESPACE_URL, "https://www.github.com/materialsintelligence/propnet/")
@@ -74,64 +64,3 @@ def references_to_bib(refs):
         parsed_refs.append(parsed_ref)
 
     return parsed_refs
-
-
-class PropnetBuilder(Builder):
-
-    DEFAULT_MATERIAL_SYMBOL_MAP = {
-    "structure": "structure",
-    "elasticity.elastic_tensor": "elastic_tensor",
-    "band_gap.search_gap.band_gap": "band_gap_pbe",
-    }
-    """
-    Basic builder for running propnet derivations on various properties
-    """
-    def __init__(self, materials, propstore, materials_symbol_map=None,
-                 criteria=None, **kwargs):
-        """
-        Args:
-            materials (MongoStore): store of materials properties
-            materials_symbol_map (dict): mapping of keys in materials
-                store docs to symbols
-            propstore (MongoStore): store of propnet properties
-            **kwargs: kwargs for builder
-        """
-        self.materials = materials
-        self.propstore = propstore
-        self.criteria = criteria
-        self.materials_symbol_map = materials_symbol_map \
-                                    or self.DEFAULT_MATERIAL_SYMBOL_MAP
-
-    def get_items(self):
-        props = list(self.materials_symbol_map.keys())
-        props += ["task_id", "pretty_formula"]
-        docs = self.materials.query(criteria=self.criteria, properties=props)
-        for doc in docs:
-            logger.info("Processing %s", doc['task_id'])
-            yield doc
-
-    def process_item(self, item):
-        # Define quantities corresponding to materials doc fields
-        # Attach quantities to materials
-        quantities = []
-        material = Material()
-        for mkey, property_name in self.materials_symbol_map.items():
-            value = get(item, mkey)
-            material.add_quantity(Quantity(property_name, value))
-
-        # Use graph to generate expanded quantity pool
-        graph = Graph()
-        graph.add_material(material)
-        graph.evaluate()
-
-        # Format document and return
-        doc = graph._symbol_to_quantity
-        doc = {symbol.name: q_list for symbol, q_list in doc.items()}
-        doc.update({"task_id": item["task_id"],
-                    "pretty_formula": item["pretty_formula"]})
-        return doc
-
-    def update_targets(self, items):
-        items = [jsanitize(item, strict=True) for item in items]
-        self.propstore.update(items)
-
